Post feature and section has been finished

+++ app/Http/Controllers/AdminUsersController.php
@@ -161,12 +161,13 @@ class AdminUsersController extends Controller
 
         $user = User::findOrFail($id);
 
-        unlink(public_path() . $user->photo->file);
+        if ($user->photo) {
+            unlink(public_path() . $user->photo->file);
+        }
 
         $user->delete();
-
-//        $request->session
-//            session()
+//      $request->session
+//      session()
         Session::flash('deleted_user', 'the user has been deleted');
         return redirect('/admin/users');
     }
